Update patch set 1
Patch Set 1:

> 1. Do you think it is realistic to get cloud account with admin
 > privileges in real world use cases. The administrator is a "god"
 > entity in the cloud without any restrictions?
Fujinhua: I agree with you about this, so in real use cases, we can config the tenant only has list tenant right.
 > 2. The list tenants is also OpenStack specific the notion of tenant
 > does not exists in Azure nor in vCloud nor in EC2.
Fujinhua: The list tenants interface of multi-cloud is different from that in openstack, it is an abstract interface of all vims, for example, vcloud can use vapp to replace tenant, here list tenant is an unified interface for all vims. In R2, multi-cloud plan to improve the adaptor interfaces to satisfy all vims.
 > 3. I am not sure that multi cloud is a really "multi cloud". The
 > current VF-C implementation only works with OpenStack and nothing
 > else even if it is using multi cloud.
Fujinhua: In Onap R1 use cases, only vio and windriver vim which are consistent with openstack interfaces, which lead to your misunderstanding to multi-cloud.
 > 4. windriver vio also supports the identity
 > 
 > Currently all clouds supported by multicloud supports the identity
 > call. Azure in multicloud is an empty directory. I think in the
 > current situation it is a lot better to use token get, than list
 > tenants.
Fujinhua: Yes, you are right about this, but the token interface is a proxy interface which are called by so and app-c components, vf-c cannot use the proxy interfaces, vf-c can only use the abstract interfaces(unified interfaces for all vims), in onap R2 or later, multi-cloud plan to discard the proxy interfaces little by little, and only support abstract interfaces.
Finally, in your codes, you set the tenantid with the tenant which is defined in ESR, not the tenant in connect info, here we need get the tenantid of the tenant in connect info, you can see the original codes again.

Patch-set: 1