Update patch set 1
Patch Set 1:

As a committer of Multi-VIM/Cloud project, I am going to judge which points are correct or not, just want to share a bigger scope of context and try to come up with some generic questions and possible solutions.

Context: multicloud is a shim layer to mediate interaction between ONAP components and underlying VIM/clouds, in ONAP R1 we implemented 2 set of NBI (northbound API) for ONAP components: Proxied OpenStack APIs which are tested with SO/APPC/DCAEgen2, and another set of NBIs which is abstract layer between VF-C and MultiCloud. The reason we had 2 seperate NBIs is "backward compatibility", since SO/APPC/DCAEgen2 seed codes are using OpenStack APIs and VF-C seed codes using the abstract APIs of MultiVIM. Now MultiVIM becomes MultiCloud. To facilitate the integration between MultiCloud with all these components, we decided exposes NBIs in this way.
On the southbound, MultiCloud has implemented adaptors (referred as  multicloud plugins) for vanilla OpenStack Ocata and Wind River's OpenStack (Titanium Cloud), as well as VMware VIO. Microsoft Azure was planning to implemented a translator from OpenStack API to Azure API, but didn't happen yet.

Questions:
Q1: who and why ONAP components need privilege to 'list project'?
Answer 1: well, it is not only that VF-C performs this operation, but also APPC does the same. Now we know exactly why VF-C need this operation, while I am still not clear why APPC need it. I had raised the question to APPC team but no clear answer yet. If we take further step to think about it in bigger scope, it is actually be the question 2 as below. 

Q2: What APIs and privileges does ONAP platform as a whole needed to underlying VIM/Clouds?
A2: There are multiple ONAP components interacting with underlying VIM/Clouds, assuming various APIs/privileges granted to them respectively. That means, just membership of a tenant/user to underlying VIM/Cloud is not enough for ONAP platform to orchestrate VNFs. On the other side, it seems 'admin' privilege is too much for an orchestrator. This is not the observation observed only from ONAP community, we had encountered the same challenge in other practices (I am from Wind River and we offered commercialized OpenStack ) . So IMHO, I think we need re-think about the role definition in underlying VIM/Clouds in context of NFV. As a experimental, we tried to add a role named 'can_list_projects' to fulfill ONAP's requirement temporarily.

Q4: How MultiCloud can help?
A4: that is case by case. e.g. given the reason why VF-C need 'list project' , multicloud could return a list of projects which was registrated in AAI instead of forwarding the requests to underlying VIM. Who and When to register those list of projects into AAI is another question and feasible, either manually or automaticatlly. 

Suggestion to this commit change:
You cannot fix/workaround the issue you intends to resolve by this single commit, we need come up with more holistic solution.
I would suggest you put a hold on this change and keep an eye on the community discussion on related topic, or just initiate this discussion with the whole community. 
Thanks.

Patch-set: 1